sdk: no need for named future for history sharing
